reorganize endpoint
