+++ src/server/host/Endpoints/User/UserGroup.cs
@@ -1,0 +1,11 @@
+using FastEndpoints;
+
+namespace MinigolfFriday.Host.Endpoints.User;
+
+public class UserGroup : Group
+{
+    public UserGroup()
+    {
+        Configure("user", x => { });
+    }
+}
