+++ src/server/host/Endpoints/User/settings/UserSettingsGroup.cs
@@ -1,15 +1,17 @@
 using FastEndpoints;
+using MinigolfFriday.Host.Endpoints.User;
 
-namespace MinigolfFriday.Host.Endpoints.UserSettings;
+namespace MinigolfFriday.Host.Endpoints.User.Settings;
 
 public class UserSettingsGroup : Group
 {
     public UserSettingsGroup()
     {
         Configure(
-            "user-settings",
+            "settings",
             x =>
             {
+                x.Group<UserGroup>();
                 x.Description(x => x.WithTags("UserSettings"));
             }
         );
